[GPU] Extract debug code from network::execute()
