+++ src/plugins/intel_gpu/src/graph/include/program_dump_graph.h
@@ -14,6 +14,6 @@ std::string get_dir_path(const ExecutionConfig& config);
 void dump_graph_optimized(std::ofstream&, const program&);
 void dump_graph_processing_order(std::ofstream&, const program&);
 void dump_graph_init(std::ofstream&, const program&,
-                     std::function<std::shared_ptr<primitive_inst>(const primitive_id&)> get_primitive_inst = nullptr);
+                     std::function<std::shared_ptr<const primitive_inst>(const primitive_id&)> get_primitive_inst = nullptr);
 void dump_graph_info(std::ofstream&, const program&);
 }  // namespace cldnn
